backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9